Duplicate React copy in the same app
Setting `React` as a dependency will result in `Invalid hook call. Hooks can only be called inside of the body of a function component.`. This happens when calling or using `useTimer` hook. Following the documentation and after verifying every part of the issue, I found that you are depending on `react` so I had to use `peerDependencies`.

For more information:

- React docs: https://reactjs.org/warnings/invalid-hook-call-warning.html#duplicate-react
- Npm docs: https://nodejs.org/es/blog/npm/peer-dependencies/